Add use strict to all FixedDataTable related files that also use ES6 classes
With our ES6 class transform if `"use strict";` is not defined at the top of the file it will insert it within each method, this causes a weird bug in Chrome when the file is `eval`ed which is common with webpack for source maps support.

I have added "use strict"; to all files that were also using ES6 classes and taken the chance to standardize some of the `"use strict";` calls.

Solves issue: #20