Allow scroll wheel to be totally disabled
Sometimes it might be useful for the scroll to be totally disabled and the wheel events to be handled by the parent. This is not the current case. Here are some quick cases where the table scroll wheel should be handled by the parent:

- Activating the table scroll only on certain events (like on click)
- Enabling table scroll only on specific layouts (fullscreen)
- Don't let the parent wheel be stopped when the mouse goes over the table (this is currently possible with 'hidden', but only on the first pass)

This is just a quick hack, im sure that a much more elegant solution is possible that can achieve this.