adjust size of table to screen size
This adds scroll bars to the table, and makes the specified width a max width.

I use this on my website; I specified the width in React code, and users with large monitors can use the very large table effectively. If they are offsite on their mobile, they can still access the table by scrolling through the rows and columns.

Perhaps this should be a configurable option; if I had a hint towards a good approach, I would be happy to implement it.