[BLT-1033] Fix columnKey warning
FixedDataTableCell clones a React element and passes some props to the clone. When a React element is a React DOM element, React will interpret passed props as HTML attributes and attempt to apply them if they are valid HTML attributes. In our case, FixedDataTableCell clones props.cell and passes the columnKey prop, but if props.cell is a 'div', React will apply columnKey as a HTML attribute and because it isn't a valid HTML attribute, it will give a warning.

This commit fixes that by deleting the columnKey prop if props.cell is a 'div'.